Call child realize from parent realize.
Previously we were calling r.g. `TreeDataGridCellsPresenter.UpdateRowIndex` from `TreeDataGridRow.Realize` instead of `TreeDataGridCellsPresenter.Realize`. This was causing the cells presenter to be stuck in a half-realized state. Make sure we call child `Realize` on realize and parent `UpdateRowIndex` only when updating row index. Added checks to ensure that `UpdateRowIndex`  can't be called on an unrealized row or cell.

Fixes #218