Don't write value to model when change comes from model.
If a changed value comes from the model, then we shouldn't be writing it back to the model. Usually this would be a no-op but when a `StringFormat` is applied it causes the formatted value to be sent back to the underlying data model, and if the `StringFormat` is not round-trippable causes an exception.